feat: progress on chat
