+++ lib/routes/Play/PlayOfflineRoute.tsx
@@ -1,16 +1,16 @@
 import React, { useContext, useEffect } from "react";
 import { previewContentEditable } from "../../components/ContentEditable/ContentEditable";
 import { PageMeta } from "../../components/PageMeta/PageMeta";
-import { Session } from "../../components/Scene/Scene";
 import { CharactersContext } from "../../contexts/CharactersContext/CharactersContext";
 import { useLogger } from "../../contexts/InjectionsContext/hooks/useLogger";
 import { SettingsContext } from "../../contexts/SettingsContext/SettingsContext";
 import { useScene } from "../../hooks/useScene/useScene";
+import { useTranslate } from "../../hooks/useTranslate/useTranslate";
+import { Session } from "./components/Session/Session";
 import {
   useSession,
   useSessionCharacterSheets,
-} from "../../hooks/useScene/useSession";
-import { useTranslate } from "../../hooks/useTranslate/useTranslate";
+} from "./components/Session/useSession";
 
 export const PlayOfflineRoute: React.FC<{}> = () => {
   const settingsManager = useContext(SettingsContext);
